Fix int overflow

+++ src/libprimer3.h
@@ -110,7 +110,7 @@ typedef enum p3_output_type {
 
 /* pr_append_str is an append-only string ADT. */
 typedef struct pr_append_str {
-  int storage_size;
+  size_t storage_size;
   char *data;
 } pr_append_str;
